Moved complexity of selecting correct lru_cache to compatibilty.py
